Fix _prevSteps int type (avoid rerendering(blinking) #bugfix

+++ src/UserInterface/Components/InfoPanel/StepCounter.h
@@ -17,7 +17,8 @@ class StepCounter {
 
 		const uint8_t POS_X = 20;
 		const uint8_t POS_Y = 170;
-		uint8_t _prevStepCount = 0;
+
+		uint _prevStepCount = 0;
 
 		void _clearDisplay();
 		void _renderIcon();
